Restored XP support for release binaries
Restored v140_xp platform toolset use, with the missing include
and library paths added in manually. Hopefully that's good enough
as a work-around until Microsoft release a proper fix.

Also fixed winusb and zlib not being correctly delay-loaded under
Windows, and cleaned up some warnings to fix Windows x64 builds.

+++ src/FluxDecoder.cpp
@@ -20,7 +20,7 @@ FluxDecoder::FluxDecoder (const std::vector<std::vector<uint32_t>> &flux_revs, i
 
 int FluxDecoder::flux_revs () const
 {
-	return m_flux_revs.size();
+	return static_cast<int>(m_flux_revs.size());
 }
 
 int FluxDecoder::flux_count () const
